requirements.txt
flask
redis
websocket-client
rel
openpyxl
